Fix Pal ID's and replace datatypes recursively in SDK

+++ decomp/SDK/include/dolphin/demo/DEMOInit.h
@@ -16,7 +16,7 @@ void DEMOSetTevColorIn(enum _GXTevStageID stage, enum _GXTevColorArg a, enum _GX
 void DEMOSetTevOp(enum _GXTevStageID id, enum _GXTevMode mode);
 struct _GXRenderModeObj* DEMOGetRenderModeObj();
 u32                      DEMOGetCurrentBuffer(void);
-void                     DEMOEnableBypassWorkaround(unsigned long timeoutFrames);
+void                     DEMOEnableBypassWorkaround(u32 timeoutFrames);
 void                     DEMOReInit(struct _GXRenderModeObj* mode);
 
 #endif // _DOLPHIN_DEMOINIT_H_
