+++ decomp/SDK/include/dolphin/demo/DEMOPuts.h
@@ -16,8 +16,8 @@ typedef enum
 
 void                 DEMOSetFontType(DMFontType attr);
 void                 DEMOLoadFont(enum _GXTexMapID texMap, enum _GXTexMtx texMtx, DMTexFlt texFlt);
-void                 DEMOSetupScrnSpc(long width, long height, float depth);
-void                 DEMOInitCaption(long font_type, long width, long height);
+void                 DEMOSetupScrnSpc(s32 width, long height, float depth);
+void                 DEMOInitCaption(s32 font_type, long width, long height);
 void                 DEMOPuts(s16 x, s16 y, s16 z, char* string);
 void                 DEMOPrintf(s16 x, s16 y, s16 z, char* fmt, ...);
 struct OSFontHeader* DEMOInitROMFont();
