+++ decomp/SDK/src/dolphin/axfx/reverb_std.c
@@ -6,19 +6,19 @@
 #include "fake_tgmath.h"
 
 // functions
-static void DLsetdelay(struct AXFX_REVSTD_DELAYLINE* dl, long lag);
-static void DLcreate(struct AXFX_REVSTD_DELAYLINE* dl, long max_length);
+static void DLsetdelay(struct AXFX_REVSTD_DELAYLINE* dl, s32 lag);
+static void DLcreate(struct AXFX_REVSTD_DELAYLINE* dl, s32 max_length);
 static void DLdelete(struct AXFX_REVSTD_DELAYLINE* dl);
 static int  ReverbSTDCreate(struct AXFX_REVSTD_WORK* rv, float coloration, float time, float mix, float damping,
                             float predelay);
 static int  ReverbSTDModify(struct AXFX_REVSTD_WORK* rv, float coloration, float time, float mix, float damping,
                             float predelay);
-static void HandleReverb(long* sptr, struct AXFX_REVSTD_WORK* rv);
-static void ReverbSTDCallback(long* left, long* right, long* surround, struct AXFX_REVSTD_WORK* rv);
+static void HandleReverb(s32* sptr, struct AXFX_REVSTD_WORK* rv);
+static void ReverbSTDCallback(s32* left, long* right, long* surround, struct AXFX_REVSTD_WORK* rv);
 static void ReverbSTDFree(struct AXFX_REVSTD_WORK* rv);
 
 static void
-DLsetdelay(struct AXFX_REVSTD_DELAYLINE* dl, long lag)
+DLsetdelay(struct AXFX_REVSTD_DELAYLINE* dl, s32 lag)
 {
     dl->outPoint = dl->inPoint - (lag * 4);
     while (dl->outPoint < 0)
@@ -28,7 +28,7 @@ DLsetdelay(struct AXFX_REVSTD_DELAYLINE* dl, long lag)
 }
 
 static void
-DLcreate(struct AXFX_REVSTD_DELAYLINE* dl, long max_length)
+DLcreate(struct AXFX_REVSTD_DELAYLINE* dl, s32 max_length)
 {
     dl->length = (max_length * 4);
     dl->inputs = OSAllocFromHeap(__OSCurrHeap, max_length * 4);
@@ -50,7 +50,7 @@ ReverbSTDCreate(struct AXFX_REVSTD_WORK* rv, float coloration, float time, float
 {
     u8          i;
     u8          k;
-    static long lens[4] = {
+    static s32 lens[4] = {
         0x000006FD,
         0x000007CF,
         0x000001B1,
@@ -150,7 +150,7 @@ const static float  value0_6 = 0.6f;
 const static double i2fMagic = 4503601774854144.0;
 
 asm static void
-HandleReverb(register long* sptr, register struct AXFX_REVSTD_WORK* rv)
+HandleReverb(register s32* sptr, register struct AXFX_REVSTD_WORK* rv)
 {
     nofralloc stwu r1, -144(r1)stmw r17, 8(r1)stfd f14, 88(r1)stfd f15, 96(r1)stfd f16, 104(r1)stfd f17,
         112(r1)stfd f18, 120(r1)stfd f19, 128(r1)stfd f20, 136(r1)lis r31, value0_3 @ha lfs f6,
@@ -553,7 +553,7 @@ HandleReverb(register long* sptr, register struct AXFX_REVSTD_WORK* rv)
 }
 
 static void
-ReverbSTDCallback(long* left, long* right, long* surround, struct AXFX_REVSTD_WORK* rv)
+ReverbSTDCallback(s32* left, long* right, long* surround, struct AXFX_REVSTD_WORK* rv)
 {
     HandleReverb(left, rv);
 }
