+++ decomp/SDK/src/dolphin/card/CARDRead.c
@@ -180,10 +180,10 @@ CARDReadAsync(CARDFileInfo* fileInfo, void* buf, s32 length, s32 offset, CARDCal
     return result;
 }
 
-long
-CARDRead(struct CARDFileInfo* fileInfo, void* buf, long length, long offset)
+s32
+CARDRead(struct CARDFileInfo* fileInfo, void* buf, s32 length, long offset)
 {
-    long result = CARDReadAsync(fileInfo, buf, length, offset, __CARDSyncCallback);
+    s32 result = CARDReadAsync(fileInfo, buf, length, offset, __CARDSyncCallback);
 
     if (result < 0)
     {
