+++ decomp/SDK/src/dolphin/dvd/dvdqueue.c
@@ -10,12 +10,12 @@ static struct
     /* 0x04 */ struct DVDCommandBlock* prev;
 } WaitingQueue[4];
 
-static struct DVDCommandBlock* PopWaitingQueuePrio(long prio);
+static struct DVDCommandBlock* PopWaitingQueuePrio(s32 prio);
 
 void
 __DVDClearWaitingQueue(void)
 {
-    unsigned long           i;
+    u32           i;
     struct DVDCommandBlock* q;
 
     for (i = 0; i < 4; i++)
@@ -27,7 +27,7 @@ __DVDClearWaitingQueue(void)
 }
 
 int
-__DVDPushWaitingQueue(long prio, struct DVDCommandBlock* block)
+__DVDPushWaitingQueue(s32 prio, struct DVDCommandBlock* block)
 {
     int                     enabled = OSDisableInterrupts();
     struct DVDCommandBlock* q = (struct DVDCommandBlock*)&WaitingQueue[prio];
@@ -41,7 +41,7 @@ __DVDPushWaitingQueue(long prio, struct DVDCommandBlock* block)
 }
 
 static struct DVDCommandBlock*
-PopWaitingQueuePrio(long prio)
+PopWaitingQueuePrio(s32 prio)
 {
     struct DVDCommandBlock* tmp;
     int                     enabled;
@@ -62,7 +62,7 @@ PopWaitingQueuePrio(long prio)
 struct DVDCommandBlock*
 __DVDPopWaitingQueue(void)
 {
-    unsigned long           i;
+    u32           i;
     int                     enabled;
     struct DVDCommandBlock* q;
 
@@ -82,7 +82,7 @@ __DVDPopWaitingQueue(void)
 int
 __DVDCheckWaitingQueue(void)
 {
-    unsigned long           i;
+    u32           i;
     int                     enabled;
     struct DVDCommandBlock* q;
 
@@ -123,7 +123,7 @@ __DVDDequeueWaitingQueue(struct DVDCommandBlock* block)
 int
 __DVDIsBlockInWaitingQueue(struct DVDCommandBlock* block)
 {
-    unsigned long           i;
+    u32           i;
     struct DVDCommandBlock* start;
     struct DVDCommandBlock* q;
 
@@ -167,7 +167,7 @@ static char* CommandNames[16] = {
 void
 DVDDumpWaitingQueue(void)
 {
-    unsigned long           i;
+    u32           i;
     struct DVDCommandBlock* start;
     struct DVDCommandBlock* q;
